requirements.txt
fastapi
boto3
uvicorn
PyYAML
requests
secure-smtplib
pytest
pytest-cov
bandit
safety